Comments about your code
Amazing work. These are the things that you're doing great!

All the test cases! 
Constantly checking for wrong inputs from the user!!
The only thing you should consider is that your midnight variable in your function `timeSinceMidnight()` is a little off because it should be 2022, 02, 03 - not January 2022, 01, 03. 
Keep doing this fantastic work!.